Update pwn_laravel.py
If the URL is valid but not the intended target then the html response is received which is not useful in the case of exploit. But since the url is valid no error's are thrown. BeautifulSoup is used to parse the response.text and check if it contains html elements, if it does then the output is saved to a log.txt file and message indicating invalid url is shown on the console.